// apps/react-frontend/src/api/client/models/company-branch-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Eazy Recycling Backend Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AddressDto } from './address-dto';

/**
 * 
 * @export
 * @interface CompanyBranchResponse
 */
export interface CompanyBranchResponse {
    /**
     * 
     * @type {string}
     * @memberof CompanyBranchResponse
     */
    'id': string;
    /**
     * 
     * @type {AddressDto}
     * @memberof CompanyBranchResponse
     */
    'address': AddressDto;
    /**
     * 
     * @type {string}
     * @memberof CompanyBranchResponse
     */
    'companyId': string;
}